Move layout types constant to the common bundle
