+++ x-pack/examples/reporting_example/public/containers/main.tsx
@@ -39,7 +39,8 @@ import type {
   JobParamsPDFV2,
   JobParamsPNGV2,
 } from '../../../../plugins/reporting/public';
-import { constants, ReportingStart } from '../../../../plugins/reporting/public';
+import { LayoutTypes } from '../../../../plugins/screenshotting/public';
+import { ReportingStart } from '../../../../plugins/reporting/public';
 
 import { REPORTING_EXAMPLE_LOCATOR_ID } from '../../common';
 
@@ -87,7 +88,7 @@ export const Main = ({ basename, reporting, screenshotMode }: ReportingExampleAp
   const getPDFJobParamsDefault = (): JobAppParamsPDF => {
     return {
       layout: {
-        id: constants.LAYOUT_TYPES.PRESERVE_LAYOUT,
+        id: LayoutTypes.PRESERVE_LAYOUT,
       },
       relativeUrls: ['/app/reportingExample#/intended-visualization'],
       objectType: 'develeloperExample',
@@ -99,7 +100,7 @@ export const Main = ({ basename, reporting, screenshotMode }: ReportingExampleAp
     return {
       version: '8.0.0',
       layout: {
-        id: constants.LAYOUT_TYPES.PRESERVE_LAYOUT,
+        id: LayoutTypes.PRESERVE_LAYOUT,
       },
       locatorParams: [
         { id: REPORTING_EXAMPLE_LOCATOR_ID, version: '0.5.0', params: { myTestState: {} } },
@@ -114,7 +115,7 @@ export const Main = ({ basename, reporting, screenshotMode }: ReportingExampleAp
     return {
       version: '8.0.0',
       layout: {
-        id: constants.LAYOUT_TYPES.PRESERVE_LAYOUT,
+        id: LayoutTypes.PRESERVE_LAYOUT,
       },
       locatorParams: {
         id: REPORTING_EXAMPLE_LOCATOR_ID,
@@ -131,7 +132,7 @@ export const Main = ({ basename, reporting, screenshotMode }: ReportingExampleAp
     return {
       version: '8.0.0',
       layout: {
-        id: constants.LAYOUT_TYPES.PRESERVE_LAYOUT,
+        id: LayoutTypes.PRESERVE_LAYOUT,
       },
       locatorParams: {
         id: REPORTING_EXAMPLE_LOCATOR_ID,
@@ -148,7 +149,7 @@ export const Main = ({ basename, reporting, screenshotMode }: ReportingExampleAp
     return {
       version: '8.0.0',
       layout: {
-        id: print ? constants.LAYOUT_TYPES.PRINT : constants.LAYOUT_TYPES.PRESERVE_LAYOUT,
+        id: print ? LayoutTypes.PRINT : LayoutTypes.PRESERVE_LAYOUT,
         dimensions: {
           // Magic numbers based on height of components not rendered on this screen :(
           height: 2400,
