+++ x-pack/plugins/screenshotting/common/index.ts
@@ -7,3 +7,4 @@
 
 export type { Context } from './context';
 export type { LayoutParams } from './layout';
+export { LayoutTypes } from './layout';
