+++ x-pack/plugins/screenshotting/server/layouts/index.ts
@@ -26,15 +26,6 @@ interface LayoutSelectors {
 
 export type Layout = BaseLayout & LayoutSelectors & Partial<Size>;
 
-/**
- * Supported layout types.
- */
-export const LayoutTypes = {
-  PRESERVE_LAYOUT: 'preserve_layout',
-  PRINT: 'print',
-  CANVAS: 'canvas', // no margins or branding in the layout
-};
-
 export const DEFAULT_SELECTORS: LayoutSelectorDictionary = {
   screenshot: '[data-shared-items-container]',
   renderComplete: '[data-shared-item]',
