+++ x-pack/plugins/screenshotting/server/layouts/preserve_layout.ts
@@ -7,7 +7,8 @@
 import path from 'path';
 import type { CustomPageSize } from 'pdfmake/interfaces';
 import type { LayoutSelectorDictionary, Size } from '../../common/layout';
-import { DEFAULT_SELECTORS, LayoutTypes } from '.';
+import { LayoutTypes } from '../../common';
+import { DEFAULT_SELECTORS } from '.';
 import type { Layout } from '.';
 import { BaseLayout } from './base_layout';
 import type { PageSizeParams } from './base_layout';
